refactor: extract `verifyWithTsParser`
